+++ @commitlint/config-conventional/index.js
@@ -1,8 +1,9 @@
 module.exports = {
+	parserPreset: 'conventional-changelog-conventionalcommits',
 	rules: {
 		'body-leading-blank': [1, 'always'],
 		'footer-leading-blank': [1, 'always'],
-		'header-max-length': [2, 'always', 72],
+		'header-max-length': [2, 'always', 100],
 		'scope-case': [2, 'always', 'lower-case'],
 		'subject-case': [
 			2,
@@ -23,6 +24,7 @@ module.exports = {
 				'docs',
 				'feat',
 				'fix',
+				'improvement',
 				'perf',
 				'refactor',
 				'revert',