+++ @commitlint/rules/src/signed-off-by.ts
@@ -1,0 +1,20 @@
+import message from '@commitlint/message';
+import toLines from '@commitlint/to-lines';
+import {Rule} from './types';
+
+export const signedOffBy: Rule<string> = (
+	parsed,
+	when = 'always',
+	value = ''
+) => {
+	const lines = toLines(parsed.raw).filter(Boolean);
+	const last = lines[lines.length - 1];
+
+	const negated = when === 'never';
+	const hasSignedOffBy = last.startsWith(value);
+
+	return [
+		negated ? !hasSignedOffBy : hasSignedOffBy,
+		message(['message', negated ? 'must not' : 'must', 'be signed off'])
+	];
+};